services/market-data/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
redis==3.5.3
httpx==0.19.0
kafka-python==2.0.2
pydantic==1.8.2
websockets==10.0 